+++ src/main/java/gift/controller/MemberController.java
@@ -1,5 +1,9 @@
-package gift;
+package gift.controller;
 
+import gift.service.JwtService;
+import gift.util.LoginMember;
+import gift.domain.Member;
+import gift.service.MemberService;
 import jakarta.validation.Valid;
 import java.util.HashMap;
 import java.util.Map;
