+++ src/main/java/gift/domain/Member.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.domain;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
