+++ src/main/java/gift/dto/WishRequest.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.dto;
 
 import jakarta.validation.constraints.NotNull;
