+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package gift;
+package gift.repository;
 
+import gift.domain.Product;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
