+++ src/main/java/gift/repository/WishRepository.java
@@ -1,5 +1,6 @@
-package gift;
+package gift.repository;
 
+import gift.domain.Wish;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
