+++ src/main/java/gift/util/LoginMember.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.util;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
